requirements.txt
pygame
PyOpenGL
numpy
scipy
gymnasium
stable-baselines3
